Refactor: View Controller - Api Controller 분리
