+++ src/main/java/com/example/springbootboard/controller/MainController.java
@@ -1,18 +1,18 @@
-package com.example.springbootboard.common;
+package com.example.springbootboard.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 
-@Controller
-@RequestMapping("/")
 @RequiredArgsConstructor
+@Controller("/")
 public class MainController {
 
     @GetMapping()
     public String getIndexPage(Model model) {
         return "index";
     }
+
+
 }
