+++ src/main/java/com/example/springbootboard/domain/comments/CommentController.java
@@ -1,16 +1,13 @@
-package com.example.springbootboard.controller;
+package com.example.springbootboard.domain.comments;
 
-import com.example.springbootboard.domain.comments.CommentService;
 import com.example.springbootboard.domain.comments.dto.CommentDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@Controller
-@RequestMapping("/boards")
 @RequiredArgsConstructor
+@Controller("/boards")
 public class CommentController {
 
     private final CommentService commentService;
