+++ src/main/java/com/example/springbootboard/domain/posts/PostController.java
@@ -2,49 +2,19 @@ package com.example.springbootboard.domain.posts;
 
 import com.example.springbootboard.domain.posts.dto.PostRequestDTO;
 import com.example.springbootboard.domain.posts.dto.PostResponseDTO;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
-@Controller
-@RequestMapping("/boards")
 @Slf4j
 @RequiredArgsConstructor
+@Controller("/boards")
 public class PostController {
     private final PostService postService;
-
-    @GetMapping("/list")
-    public String getPostList(Model model, @RequestParam String teamName, @RequestParam(required = false, defaultValue = "") String search, Pageable pageable) {
-        Page<PostResponseDTO> postResponseDTOList = null;
-
-        if (teamName.equals("all")) {
-            postResponseDTOList = postService.findAllWithPagination(search, pageable);
-        } else {
-            postResponseDTOList = postService.findAllByTeamNameWithPagination(teamName, search, pageable);
-        }
-
-        model.addAttribute("listName", "Post");
-        model.addAttribute("responseDTOList", postResponseDTOList.toList());
-        model.addAttribute("totalCount", postResponseDTOList.getTotalElements());
-
-        return "boards/postList";
-    }
-
-    @GetMapping("/write")
-    public String getWritePostPage(Model model) {
-        return "boards/writePostPage";
-    }
 
     @PostMapping("/write")
     public String postWritePostPage(Model model, PostRequestDTO postRequestDTO, MultipartFile[] multipartFiles) {
@@ -58,13 +28,6 @@ public class PostController {
         return "redirect:detail?teamName=" + postRequestDTO.getTeamName() + "&postId=" + postResponseDTO.getPostId();
     }
 
-    @GetMapping("/update")
-    public String getUpdatePostPage(Model model, @RequestParam Long postId) {
-        PostResponseDTO postResponseDTO = postService.findById(postId);
-        model.addAttribute(postResponseDTO);
-
-        return "boards/updatePostPage";
-    }
 
     @PostMapping("/update")
     public String postUpdatePostPage(Model model, PostRequestDTO postRequestDTO) {
@@ -77,29 +40,6 @@ public class PostController {
         return "redirect:detail?teamName=" + postResponseDTO.getTeamName() + "&postId=" + postRequestDTO.getPostId();
     }
 
-    @GetMapping("/detail")
-    @ApiOperation(value = "상세 페이지", notes = "")
-    @ApiResponses({
-            @ApiResponse(code = 200, message = "성공"),
-            @ApiResponse(code = 401, message = "인증 실패"),
-            @ApiResponse(code = 404, message = "사용자 없음"),
-            @ApiResponse(code = 500, message = "서버 오류")
-    })
-    public String getDetailPage(Model model, @RequestParam Long postId, @RequestParam String teamName) {
-        PostResponseDTO postResponseDTO = postService.findById(postId);
-        postResponseDTO.setView(postService.incrementViewCount(postId));
-
-        model.addAttribute(postResponseDTO);
-        return "boards/detailPage";
-    }
-
-    @GetMapping("/delete")
-    public String getDeletePage(Model model, @RequestParam Long postId, @RequestParam String teamName) {
-        model.addAttribute("postId", postId);
-        model.addAttribute("teamName", teamName);
-
-        return "boards/deletePage";
-    }
 
     @PostMapping("/delete")
     public String deletePost(Model model, @RequestParam Long postId, @RequestParam String teamName) {
