+++ src/main/java/com/example/springbootboard/domain/uploadfiles/UploadFileController.java
@@ -1,7 +1,6 @@
-package com.example.springbootboard.controller.Rest.v1;
+package com.example.springbootboard.domain.uploadfiles;
 
 import com.amazonaws.services.s3.model.S3ObjectSummary;
-import com.example.springbootboard.domain.uploadfiles.AwsS3ServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,7 +13,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/rest/v1/aws/s3")
-public class UploadController {
+public class UploadFileController {
     @Autowired
     private AwsS3ServiceImpl awsS3Service;
